% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsToFrame.R
\name{countsToFrame}
\alias{countsToFrame}
\title{A function for converting a table of word frequencies to a data frame}
\usage{
countsToFrame(tbl, wordColName = "Word", freqColName = "Freq")
}
\arguments{
\item{tbl}{The frequency table of words}

\item{wordColName}{What to name the column containing the words}

\item{freqColName}{What to name the column containing the frequencies}
}
\value{
A data frame containing the frequency data
}
\description{
A function for converting a table of word frequencies to a data frame
}
