% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeWords.R
\name{visualizeWords}
\alias{visualizeWords}
\title{A function for visualizing word frequencies}
\usage{
visualizeWords(freqData)
}
\arguments{
\item{freqData}{A data frame of frequencies (see wordcount, countsToFrame)}
}
\value{
A ggplot plot visualizing these frequencies
}
\description{
A function for visualizing word frequencies
}
