% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcount.R
\name{wordcount}
\alias{wordcount}
\title{A function for counting the frequencies of words in a vector of strings}
\usage{
wordcount(
  data,
  n = 10,
  wordRegex = "[^ ,'\\\\[\\\\]]*",
  ignoreWords = c(),
  removePunct = FALSE,
  removeNums = FALSE,
  punctRegex = "[[:punct:]]",
  numsRegex = "[0-9]"
)
}
\arguments{
\item{data}{A vector of strings or a single string}

\item{n}{The number of top occurring words to display}

\item{wordRegex}{An optional regex to split the string apart}

\item{ignoreWords}{A vector of words to exclude from the analysis}

\item{removePunct}{Remove punctuation?}

\item{removeNums}{Remove numbers?}

\item{punctRegex}{Regex for punctuation to remove}

\item{numsRegex}{Regex for numbers to remove}
}
\value{
A sorted (descending) table of words and their frequencies
}
\description{
This function allows you to quickly analyze a large set of unorganized
strings in a customizable fashion
}
